requirements.txt
kubernetes
azure-mgmt-compute
azure-identity